wip troops
notes: rigBody3d seems to be a wrong choice for the troops, as it
crazily follows the world physics and in this game case we need to force
the troops position and have full control over it.
So, Im going to change it to characterBody3d (Probably) and try to get
every territory base location and spawn new troops based on this
territory.

Also, check how characterBody3d can move along the planet if its physics
is different from rigidbody3d

good luck tomorrow me